Butterfly/producer/gitlab/requirements.txt
kafka-python
Flask
requests